+++ CHANGES.rst
@@ -6,10 +6,12 @@ Version 0.2.0
 
 *Unreleased*
 
-- Add Field aliases (#14)
+- Support conversion between TOML, YAML (`#16`_)
+- Add Field aliases (`#14`_)
 - Support arbitrary serializers and deserializers (`#6`_)
 - General internal enhancements (`#5`_)
 
+.. _#16: https://github.com/rossmacarthur/serde/pull/16
 .. _#14: https://github.com/rossmacarthur/serde/pull/14
 .. _#6: https://github.com/rossmacarthur/serde/pull/6
 .. _#5: https://github.com/rossmacarthur/serde/pull/5